make word split split on periods too
I wanted to search for a domain, but it never matched, because the
needles was in the middle of a domain. By splitting on periods, that
part of the domain haystack becomes a separate word.

Also, I don't really feel safe modifying this code any more. I should
rewrite it in Rust.